(maint) Move Puppetfile::DSL to new location
