+++ lib/r10k/puppetfile.rb
@@ -4,6 +4,7 @@ require 'r10k/module'
 require 'r10k/util/purgeable'
 require 'r10k/errors'
 require 'r10k/content_synchronizer'
+require 'r10k/module_loader/puppetfile/dsl'
 
 module R10K
 class Puppetfile
@@ -91,7 +92,7 @@ class Puppetfile
   def load!(default_branch_override = nil)
     @default_branch_override = default_branch_override
 
-    dsl = R10K::Puppetfile::DSL.new(self)
+    dsl = R10K::ModuleLoader::Puppetfile::DSL.new(self)
     dsl.instance_eval(puppetfile_contents, @puppetfile_path)
 
     validate_no_duplicate_names(@modules)
@@ -254,36 +255,6 @@ class Puppetfile
     Pathname.new(basedir).cleanpath.to_s
   end
 
-  class DSL
-    # A barebones implementation of the Puppetfile DSL
-    #
-    # @api private
-
-    def initialize(librarian)
-      @librarian = librarian
-    end
-
-    def mod(name, args = nil)
-      if args.is_a?(Hash)
-        opts = args
-      else
-        opts = { version: args }
-      end
-
-      @librarian.add_module(name, opts)
-    end
-
-    def forge(location)
-      @librarian.set_forge(location)
-    end
-
-    def moduledir(location)
-      @librarian.set_moduledir(location)
-    end
-
-    def method_missing(method, *args)
-      raise NoMethodError, _("unrecognized declaration '%{method}'") % {method: method}
-    end
-  end
+  DSL = R10K::ModuleLoader::Puppetfile::DSL
 end
 end
